bot/requirements.txt
aiogram==3.21.0
aiohttp==3.12.13
email_validator==2.2.0
fastapi==0.116.1
pydantic==2.11.7
python-dotenv==1.1.1
redis==6.2.0
uvicorn==0.35.0
